requirements.txt
# Dubai Real Estate RAG Chat System - Dependencies
# 
# This file serves as the main requirements file for the project.
# For development, use: pip install -r requirements.txt
# For production, use: pip install -r backend/requirements.txt
#
# The backend/requirements.txt contains the complete dependency list
# including all ML libraries and advanced features.

# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic>=2.5.0
python-dotenv==1.0.0

# AI & ML
google-generativeai==0.3.2
chromadb>=0.5.0

# Data Processing
pandas>=2.0.0
numpy==1.24.3
python-multipart==0.0.6

# Document Processing
python-docx==1.1.0
PyPDF2==3.0.1
openpyxl==3.1.2
pdfplumber==0.9.0

# Monitoring & Observability
prometheus-client==0.19.0
psutil==5.9.6
sentry-sdk==1.38.0

# Performance & Caching
redis==5.0.1
asyncpg==0.29.0
aiohttp==3.9.1

# Security & Authentication
bcrypt==4.0.1
PyJWT==2.8.0
email-validator==2.0.0
cryptography==41.0.7

# HTTP & API
requests==2.31.0
tenacity==8.2.3

# Scheduling
APScheduler==3.10.4

# Utilities
werkzeug==2.3.7

# For complete dependencies including ML libraries, see:
# backend/requirements.txt
